requirements.txt
matplotlib==3.5.2
requests==2.27.1
Flask==2.3.2
Flask-Session==0.5.0
msal==1.22.0
mysql-connector-python==8.0.33
plotly==5.14.1
Werkzeug==2.3.4